omnibrowser-redix-memory-kit/server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
qdrant-client==1.7.3
redis==5.0.1
pydantic==1.10.13
python-jose[cryptography]==3.3.0
sentence-transformers==2.2.2
optimum[onnxruntime]==1.16.1
onnxruntime==1.16.3
numpy==1.26.4
orjson==3.10.0

